+++ Sources/Vapor/HTTP/Server/Application+HTTP+Server.swift
@@ -11,16 +11,8 @@ extension Application.HTTP {
                 self.application.storage[ConfigurationKey.self] ?? .init()
             }
             nonmutating set {
-                if self.application.storage.contains(Key.self) {
-                    self.application.logger.warning("Cannot modify server configuration after server has been used")
-                } else {
-                    self.application.storage[ConfigurationKey.self] = newValue
-                }
+                self.application.storage[ConfigurationKey.self] = newValue
             }
-        }
-
-        struct Key: StorageKey, LockKey {
-            typealias Value = HTTPClient
         }
 
         struct ConfigurationKey: StorageKey {
